requirements.txt
# ===========================================
# E-Learning Backend Requirements
# ===========================================
# Complete package list for Udemy-like platform
# Run: pip install -r requirements.txt

# ===========================================
# Django Core Framework
# ===========================================
Django==5.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# ===========================================
# Database Drivers
# ===========================================
# MySQL Support
PyMySQL==1.1.0
# PostgreSQL Support (for production)
# psycopg2-binary==2.9.9

# ===========================================
# Image & Media Processing
# ===========================================
Pillow>=10.2.0

# ===========================================
# Video Processing & Streaming
# ===========================================
ffmpeg-python==0.2.0

# ===========================================
# Payment Processing
# ===========================================
razorpay==1.4.1

# ===========================================
# Django Add-ons & Extensions
# ===========================================
django-cors-headers==4.3.1
django-filter==23.5.0

# ===========================================
# Environment & Configuration
# ===========================================
python-dotenv==1.0.0

# ===========================================
# Utilities & Libraries
# ===========================================
pytz==2024.1
requests==2.31.0

# ===========================================
# Development & Testing (Optional)
# ===========================================
# Uncomment for development:
# django-debug-toolbar==4.2.0
# pytest-django==4.5.2
# coverage==7.3.2
# ipdb==0.13.13

# ===========================================
# Async Task Queue
# ===========================================
celery>=5.3.0,<6
redis>=5.0.0,<6

# ===========================================
# Production (Optional)
# ===========================================
# Uncomment for production:
# gunicorn==21.2.0
# whitenoise==6.6.0
# django-redis==5.4.0
# boto3==1.34.34
# django-storages==1.14.2
# sentry-sdk==1.38.0